Fix bug when minor version number goes beyond 9
The upstream version determination might pick the wrong version if
a minor version was a prefix of another minor version. Observed when
going from EGit 5.9 to 5.10: the build would pick a JGit 5.1.* version
instead of 5.10.0 as upstream version.

Fix by including the period after the minor version in the prefix that
we check against our own version. Since "5.1." is not a prefix of
"5.10.", this will now pick the correct upstream version.

Change-Id: I0d8af8f938ac9f56448679badd1557de3d2dd8de
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>