Add a "noTests" config parameter
Add a new configuration parameter to skip tests and static checks.
If true, also skip reporting. By default false.

This can be used to make builds do only the bare minimum to produce
JARs and a p2 update site.

This is intended for exceptional use only, for instance if the CI
infrastructure is overloaded or otherwise has problems, and you need
to get a release out of the door. It is assumed that tests have been
run locally.

Change-Id: I120ab054c625de48ef67a041e70b9be6901b3b08
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>