Update patch set 2
Patch Set 2:

Eh! Thanks for merging, but I'm still running my thomas-test build to verify it's correct!

I'll update the configs on Jenkins to add the job parameter to all jobs. Usage would be:

1. Kill and ignore any verify builds.
2. Set verified+1 manually and merge.
3. Kill any triggered stable builds.
4. Start one stable build manually, setting the parameter.

Note that in (3) in EGit and EGit-Github, we currently get one job per merged change when a whole series is submitted. That's bug 552963, but I'll deal with that only after 5.12 is over.

Patch-set: 2