Do not pin the Java version used to run maven
The EGit and EGit-Github maven builds have been made to be able to
use any Java >= 1.8 for running the maven build. They will execute
the tests on a Java version determined by the BREE specified in the
MANIFEST.MF files (currently JavaSE-1.8).

The mechanism relies on the CI infrastructure providing a maven
toolchain definition for at least that BREE.

Change-Id: Ie83f986e54ae11cef5ce363d404ddf137eccf714
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>