Do not pin the Java version used to run maven for EGit
The EGit maven build has been made to be able to use any Java >= 1.8
for running the maven build. It will execute the tests on a Java
version determined by the BREE specified in the MANIFEST.MF files
(currently JavaSE-1.8).

The mechanism relies on the CI infrastructure providing a maven
toolchain definition for at least that BREE.

Change-Id: Ie83f986e54ae11cef5ce363d404ddf137eccf714
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>